Delete more bogus getSIMDStructFromField code
SIMD/HWINTRINSIC nodes end up under ADDR only when they're used as call args, thanks to impNormStructVal. That is not relevant to fgMorphCombineSIMDFieldAssignments, this kind of coalescing only works on assignments and not when all the fields of a Vector4 get passed on the stack in a call.

Besides, areArgumentsContiguous wouldn't recognize FIELD(ADDR(SIMD...)) as contiguous.