Remove unexpected creation of primitive struct typed locals
Attempting to create a struct local having primitive type (e.g. System.Int32) works but results in the promotion of the local, wasting a local variable and possibly causing other issues.

Also, the copy destination can be changed to a LCL_VAR, keeping the indir is not necessary and may affect CQ.