Cleanup IsLocalAddrExpr and return the offset too
Also change fgMorphCopyBlock's self copy check to use the offset rather than the field sequence, there's a higher chance that the offset is actually correct and it's anyway irrelevant if the field sequence matches or not.

Also check that local offset is 0 when promoting, at least for the sake of clarity. It's not readily obvious why weird code like `*(Foo*)((byte*)&foo + 1) = otherFoo;` doesn't trigger destination promotion with the current checks.