Fix lvSingleDef marking
optAddCopiesCallback only recognizes direct local definitions like ASG(LCL_VAR, x) so it's not useful to add a copy if the definition is something like BLK(ADDR(LCL_VAR)).

fgMorphOneAsgBlockOp normally simplifies full indirect definitions of scalar variables to direct definitions but that should be considered an optimization and not a transform required for correctness.