Stop using fgMorphBlockOperand in fgMorphInitBlock
Doesn't do much, if anything. It is always called with asgType = tree->GetType() and isBlockRequired = true so the only potentially useful transform it does is to simplify OBJ(ADDR(LCL_VAR)) to LCL_VAR when the OBJ and LCL_VAR have the same size.

But the variable would still end up DNER_BlockOp so this doesn't achieve much. And eventually such a transform should probably be done in LocalAddressVisitor.