Relax multireg arg OBJ - LCL_VAR type matching
There's no real need to require that the LCL_VAR under OBJ(ADDR(LCL_VAR)) have exactly the same type as the OBJ. The only necessary restriction is for the local to be at least as large as the indirection to avoid generating LCL_FLDs that read past the end of the local.

Exact type equality causes CQ issues because the JIT/VM manage to mix `C<T>` and `C<Canon>` when inlining. Such types have different class handles but they have the same layout and should be treated as the same type for most purposes, including ABI decisions.