Delete a gajillion of pointless asserts
It's really not relevant if the source local variable has a layout similar to the call arg layout. Whatever data is available in the local will be loaded using suitable typed LCL_FLDs (or in the fortunate case - using LCL_VARs of promoted struct fields) and that's all there is to it.