Delete useless code
fgMorphCast already handles the case of floating point to small int cast so there's no reason to do this again in lowering.

If anything, this could probably be handled entirely in codegen for the sake of simplicity but that may result in some optimizations being missed (e.g. CSEing the FP to INT cast when another one already exists). But doing it in lowering doesn't seem to have any advantages.