Delete weird ARM64 genPutStructArgStk reg handling
There's no way for srcAddrBaseReg to be the same as tempReg, BuildPutArgStk builds the temp defs before the src use.

For power of 2 struct sizes, that can be loaded with a singler ldr/ldp, we could delay the temp defs to avoid allocating an unnecessary temp register but that's not implemented now and given the relatively large number of available registers it seems unlikely to be worth it.