Relax arg reg/field reg type matching
Type size matching is too weak as it results in broken IR if FLOAT/INT reinterpretation occurs. Actual type matching is too strong because it rejects the SIMD8/DOUBLE case that occurs on linux-x64.

Ultimately what matters is if the field can be loaded in the register allocated to the argument. Of course, combined with reinterpretation, this allows weird cases like a SIMD16 field being passed in a FLOAT arg reg but ABIs do not impose any rules regarding the unused bits of a register used to pass structs.