Delete bogus HFA code
fgMorphMultiregStructArgs attempts to change the type of some HFA args  to the HFA element type. And the falls through to fgMorphMultiregStructArg which promptly asserts because the arg type isn't struct.

Besides, fgMorphMultiregStructArgs does this only if the struct size is equal to the HFA element type size. That is - for single element HFAs. And single element HFAs aren't passed in multiple registers...