Delete useless temp retyping code
When EvalArgsToTemps the args have the "correct" type already. fgMorphArgs has either retyped args (such as struct args passed in a single register) or expects the 2nd pass to do it. There's no reason for EvalArgsToTemps to do its own arg retyping.

This is probably a leftover from when temps were used for struct args having "odd" sizes (3, 5, 6 etc.). Then fgMorphArgs was requesting a temp and expected it to have primitive type.