QMARK can't appear in call args
fgExpandQmarkNodes ensures that QMARK appears only at the "top-level".

And if it appears then spilling it to a temp doesn't solve much, if anything.  At the very least, all previous args need to be spilled to temps as well since they'll typically need to be evaluated inside the QMARK's generated condition block. But that should be the job of the code that expands QMARKs, it's not that different from having a QMARK as the second operand of a non-commutative binary operator.