Delete useless regArgList
It is only used in an useless assert - if it's a late arg "result" is obtained from the late use so "result == use.GetNode()" always holds.

What assertion propagation or other arg updates may get wrong is to drop the GTF_LATE_ARG accidentally. The assert doesn't protect against that because it doesn't even run, the function returns earlier if GTF_LATE_ARG isn't present.

Ultimately it's safer to always use GetArgInfoByArgNum without checking GTF_LATE_ARG, at the cost of doing a linear search for argNum 0 if the this arg gets moved to a different position by SortArgs.