Restore layout size assert
The assert got moved from ClassLayout::InitializeGCPtrs to gtNewCpObjNode but gtNewCpObjNode is rarely used in the JIT so having the assert here is rather useless. Could be moved to codegen but then having GC pointer containing objects that have "odd" sizes is nonsense. The only reason the assert ever fired in ClassLayout::InitializeGCPtrs is that VM's getHeapClassSize returns an incorrect object size.