Change the last verRaiseVerifyException to BADCODE
This is somewhat strange as ECMA-335 doesn't clearly state that having a non empty stack after ENDFILTER is invalid, only that it is not verifable. But then it's not clear how it could possibily be valid (automatically empty the stack?) and anyway the JIT throws even if verification is not enabled.