Extract SIMD address pop handling
It's only needed in a few specific places and it doesn't require struct normalization. Which doesn't work properly due to it relying on incorrect importer code that has been removed - with verification disabled LDLOCA import pushed the struct type of the local on the stack, instead of pushing the byref type.