Disable struct normalization if structHandle isn't specified
When structHandle is null impSIMDPopStack attempts to retrieve it from the importer stack. But:
* normally the type info on the stack will have the same handle as the tree
* if the struct handle isn't specified then the caller doesn't really care about it, it only cares about the SIMD type

In general, the whole thing is dubious because impNormStructVal is normally called from impPopCallArgs because that's the only place where the struct handle matters.

Also, impNormStructVal wraps all LCL_VAR|FLD in OBJ and there's no reason to do this when the popped tree is used by an intrinsic.