Relax dubious assert
The JIT doesn't really care if the destination and the source of an assigment have different struct handle so it's not clear why the assert exists. Perhaps in some cases it really should be a BADCODE, e.g. stobj with different struct handle on stack is invalid IL.

Also, the JIT and VM sometimes manage to come up with different struct handles for what's really the same type (C<T> and C<Canon>) so such a check will not always work correctly.

In any case, for SIMD types it only needs to check the SIMD type itself and ignore the struct handle.