Delete bogus "zero offset" field sequence
Array data is never at offset 0 so there should always be a constant offset to attach the FirstElem field sequence.

It's possible to get a 0 constant offset if the index is negative and exactly cancels out the array data offset (e.g. byte_array[-16] on 64 bit) but such an index would lead to an IndexOutOfRangeException so it's not interesting to recognize it in VN. And even if the offset is 0 the constant is not normally removed because the is part of an addtion between REF and 0 which the JIT doesn't remove.

Besides, GenTree::ParseArrayAddress doesn't check for a "zero offset" field sequence...