Delete pointless ConstantIndexPseudoField
The only thing it's actually used for is an "not a field" assert in ParseArrayAddress. The assert itself is neither useful nor correct:
* The lack of ADDR(IND(...)) folding in the presence of GTF_IND_ARR_INDEX prevents struct field offsets from appearing in an array element address tree. In turn this makes all ParseArrayAddress field sequence building and checking useless, the result is always the null field sequence which indicates that the entire element is being accessed.
* If it weren't for the above problem then NotAField could appear in an array field sequence - if the address of a struct field in an array element is taken and then an arbitrary offset is added to it then the resulting field sequence should be NotAField. And then an assert is not appropiate, instead the access should be considered to alias the entire array (actually all arrays of the same type).
* The presence of ConstantIndexPseudoField is also the result of NotAField being set on all CNS_INT nodes, even when they do no represent offsets.

It kind of makes sense to combine ConstantIndexPseudoField and FirstElemPseudoField when the array data offset constant gets folded together with a constant index expression and thus becomes different from the offset of the first array element. But then FirstElemPseudoField is pretty much useless. It could be deleted but perhaps it can be reused to get rid of the array info map and GTF_IND_ARR_INDEX...