Don't generate primitive type block copies
CPOBJ is normally used to copy struct types but nothing prevents it from being used to copy primitive types as well.

However, it doesn't make sense to import a primtive type CPOBJ as a block copy - doing so can generated weird IR where the assignment LHS is a BLK of STRUCT type but with a primitive type layout. And the RHS can be a primitive type too, thanks to impImportCpObj blindly eliminating ADDRs without checking the type of the source location.

Luckily it doesn't matter too much as the C# compiler doesn't  seem to ever generate CPOBJ. VC++ uses it, usually to copy struct values when  addresses are already available (e.g. between 2 fields of a reference type but not between 2 locals).

Some IL tests also use it, for both struct and primitive type copying.