optIsFieldAddr doesn't need to be recursive
Morph should have already colapsed a sequence of field accesses to a single
ADD so we're really looking for a single ADD(ref, offset [field seq]),
or a zero offset field sequence (sometimes attached to helpers that return
static field addresses). Attempting to handle successive ADD nodes should
not be necessary and is kind of risky too - if some kind of expression
optimization reassociates the expression then the field sequence will have
an incorect order.