Remove dubios zero offset field seq transfer
It's the caller's responsibility to set a proper field sequence on the LCL_FLD. ChangeOper has no way of knowing what the original node was and what the field sequence meant. Usually a zero field sequence is attached to a node that produces an address, that's not the same thing as LCL_FLD's own field sequence.