Simplify JTRUE side effect handling
We don't really need to produce multiple statements if we get a COMMA tree of side effects. Though it's kind of cleaner to remove COMMAs when possible, it's more trouble than it's worth and there are already other places that don't bother to do this. Anyway at this point we're close to LIR, when COMMAs will be removed anyway.